test: add comprehensive test coverage for webapi and infra layers (#5)
* feat: Enhance API documentation and implement rate limiting

- Updated OpenAPI specification to include detailed descriptions for rate limiting and response examples for various endpoints.
- Changed response codes for account creation from 200 to 201 and added examples for success and error responses.
- Refactored account deposit and withdrawal endpoints to include detailed response examples and error handling for rate limits.
- Added tests for user and transaction creation, ensuring proper error handling and response validation.
- Implemented rate limiting middleware in the web API, allowing a maximum of 5 requests per second per IP.
- Created unit tests for rate limiting functionality to ensure correct behavior under load.
- Removed unnecessary parameters from user deletion function and updated related tests.

* feat(cli): add verbose flag to control log output

Add -v flag to enable/disable verbose logging. When disabled, log output is discarded to reduce noise in normal operation.

* style(cli): update welcome message with ascii art logo

* feat(cli): add colored output for better user experience

Add color to CLI output using fatih/color package to improve readability and user experience. Error messages are now red, success messages are blue, prompts are green, and the banner is cyan. Also added balance formatting to show 2 decimal places.

* feat(database): configure gorm logger to silent mode

Add logger configuration to silence gorm logs by default. This reduces noise in application logs while maintaining transaction skipping behavior.

fix(cli): use parsed account ID for balance check

Correct balance check to use parsed account ID instead of account object's ID to ensure consistency in account operations.

* test: centralize test logging setup in main_test.go

Move TestMain logging configuration from account_test.go to main_test.go to avoid duplication and ensure consistent logging behavior across all tests

* feat(api): add handler for fiber app initialization

Initialize fiber application with database connection and unit of work setup.
Add AWS Lambda and Vercel bridge dependencies for serverless deployment.

* test: add comprehensive test coverage for webapi and infra layers

- Add tests for auth, user, and account routes in webapi layer
- Implement test cases for repository and UoW in infra layer
- Update mockery configuration to generate mocks for service layer
- Improve test utilities with additional mock expectations
- Update host URL in requests.http for testing

* feat(auth): implement auth strategy pattern and refactor auth service

Refactor auth service to use strategy pattern for different authentication methods.
Move mock generation to internal/fixtures and update tests accordingly.
Add context support for current user ID retrieval.
Update webapi routes to use new auth strategy interface.

* test(account): add auth header to balance test request

* test: remove redundant mock assertions and set expectations once

The redundant defer assertions in SetupTestApp were removed as they're not needed. Mock expectations in getTestToken are now set to run only once using .Once() for more precise test behavior.

* refactor(auth): restructure auth strategy and service for better maintainability

- Move GetCurrentUserID logic to strategy interface
- Update auth service to use strategy pattern consistently
- Modify route handlers to accept auth strategy parameter
- Update test setup to use new auth strategy

* docs: update readme with comprehensive project details

- Expand overview with vision and key features
- Add detailed getting started instructions
- Document API endpoints and project structure
- Explain infrastructure choices and testing approach
- Include contribution guidelines and license info

* docs: update README formatting and example syntax

* refactor(test): move benchmark tests to separate files and add parallel execution

Move benchmark tests from various test files to dedicated *_benchmark_test.go files
Add t.Parallel() to test functions to enable parallel test execution

* ci: update workflow to use ubuntu-24.04-arm runner

* docs: add usage examples to readme

* test: add mock assertions cleanup in test setup

Ensure all mock expectations are verified by adding defer statements for AssertExpectations in SetupTestApp. Also remove redundant .Once() calls from mock expectations as they're unnecessary when using AssertExpectations.

* test: adjust test parallelization and mock expectations

- Remove t.Parallel() from webapi tests to avoid potential race conditions
- Add t.Parallel() to domain and service tests for better test isolation
- Update mock expectations in testutils.go to use .Maybe() for more flexible test setup

+++ pkg/domain/account.go
@@ -89,7 +89,7 @@ func (a *Account) Deposit(userID uuid.UUID, amount float64) (*Transaction, error
 	parsedAmount := int64(amount * 100) // Convert to cents for precision
 
 	// Check for overflow after conversion as well
-	if a.Balance > math.MaxInt64-parsedAmount {
+	if a.Balance > int64(math.MaxInt64)-parsedAmount {
 		return nil, ErrDepositAmountExceedsMaxSafeInt
 	}
 	slog.Info("Depositing amount", slog.Int64("amount", parsedAmount))
